+++ software/utils.c
@@ -19,9 +19,9 @@ char byte_to_printable(uint8_t byte_in){
 }
 
 // Compare two buffers. If they are the same, return 1
-uint8_t bufncmp(uint8_t *b1, uint8_t *b2, uint16_t len){
-    uint16_t samechars = 0;
-    for(uint8_t i = 0; i < len; i++){
+uint8_t bufncmp(uint8_t *b1, uint8_t *b2, uint32_t len){
+    uint32_t samechars = 0;
+    for(uint32_t i = 0; i < len; i++){
         if(b1[i] == b2[i]){
             samechars++;
         }
